fixed scratch pad reducer
