+++ frontend/src/css/desktop.css
@@ -16,8 +16,7 @@
 	align-items: center;
 	padding-top: 20px;
 	padding-bottom: 20px;
-	color: white;
-	text-shadow: 0px 0px 3px rgba(0, 0, 0, 0.5);
+	color: black;
 }
 
 .desktop-greeting {
